pyproject.toml
[project]
name = "social-media-impact-on-teen-mental-health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "polars-lts-cpu>=1.33.1",
    "pyarrow>=23.0.1",
    "setuptools<82",
    "sweetviz>=2.3.3",
    "ydata-profiling==4.6.4",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]
